# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Parking Reservation API",
        "description": "API for managing parking spot reservations, including booking, cancellation, and retrieving user reservations.",
        "termsOfService": "http://example.com/terms",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/parkings": {
            "get": {
                "tags": [
                    "Parking"
                ],
                "summary": "Get all parkings",
                "operationId": "8f7c807501d1c0e5fe7273b66e5e838b",
                "responses": {
                    "200": {
                        "description": "List of all parkings"
                    }
                }
            },
            "post": {
                "tags": [
                    "Parking"
                ],
                "summary": "Create a new parking",
                "operationId": "bec4db4758c912f3eed2404dced269a2",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "location",
                                    "capacity"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Central Parking"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Downtown, City"
                                    },
                                    "capacity": {
                                        "type": "integer",
                                        "example": 100
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Parking created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/parkings/{id}": {
            "get": {
                "tags": [
                    "Parking"
                ],
                "summary": "Get a parking by ID",
                "operationId": "9ee165016e82d3ac7a703a548cbf54ca",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Parking details"
                    },
                    "404": {
                        "description": "Parking not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Parking"
                ],
                "summary": "Update an existing parking",
                "operationId": "075630bf013386e46f257884d91d2579",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "location",
                                    "capacity"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Updated Parking Name"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "New Location"
                                    },
                                    "capacity": {
                                        "type": "integer",
                                        "example": 150
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Parking updated successfully"
                    },
                    "404": {
                        "description": "Parking not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Parking"
                ],
                "summary": "Delete a parking",
                "operationId": "abcefd5627338c4421d5cade45760c0d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Parking deleted successfully"
                    },
                    "404": {
                        "description": "Parking not found"
                    }
                }
            }
        },
        "/api/parkings/search": {
            "get": {
                "tags": [
                    "Parking"
                ],
                "summary": "Search parkings by location and radius",
                "operationId": "171d39878b9729482dd8782e6277f27c",
                "parameters": [
                    {
                        "name": "latitude",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        },
                        "example": 34.0522
                    },
                    {
                        "name": "longitude",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float"
                        },
                        "example": -118.2437
                    },
                    {
                        "name": "radius",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 5
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of parkings found"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
...
    },
...
}